legacy/requirements.txt
pydot==1.4.2
matplotlib==3.2.2
netgraph==4.11.7
networkx==2.5
plotly==5.12.0
Pillow==9.4.0
pandas==1.5.3
wandb==0.13.10
jaxtyping==0.2.14
optax==0.1.4
jaxopt==0.6
ott-jax==0.4.0